// frontend/src/services/generated/models/author.ts
/**
 * Generated by orval v7.2.0 🍺
 * Do not edit manually.
 * DOCUMENTATION
 * OpenAPI spec version: 1.0.0
 */
import type { AuthorArticlesItem } from './authorArticlesItem';
import type { AuthorAvatar } from './authorAvatar';
import type { AuthorCreatedBy } from './authorCreatedBy';
import type { AuthorLocalizationsItem } from './authorLocalizationsItem';
import type { AuthorUpdatedBy } from './authorUpdatedBy';

export interface Author {
  articles?: AuthorArticlesItem[];
  avatar?: AuthorAvatar;
  createdAt?: string;
  createdBy?: AuthorCreatedBy;
  documentId?: string;
  email?: string;
  id?: number;
  locale?: string;
  localizations?: AuthorLocalizationsItem[];
  name?: string;
  publishedAt?: string;
  updatedAt?: string;
  updatedBy?: AuthorUpdatedBy;
}